support for a NEXT token
Usually Dist::Zilla projects have a Changes file in
the repository that has a {{$NEXT}} placeholder for
the next version.  To support that, I've shoe-horned
the possibility to pass a 'next_token' argument to
the CPAN::Changes constructor. See the test file
for an example of how to use it.