add the option to sort groups
By the default, the current ordering (alphanumerical)
is kept, but I'm adding the possibility to provide
a different sorting function.

An example where this comes handy is when groups are
not all of equal importance. In my modules, I want
the 'API CHANGES' to be shown first, then
the 'ENHANCEMENTS' and then the 'BUG FIXES'.